Checkpoint as of first successful tp connection, before posix, & before mckusick starts incorporating gnodes
SCCS-vsn: sys/netiso/argo_debug.h 7.2
SCCS-vsn: sys/netiso/clnl.h 7.2
SCCS-vsn: sys/netiso/clnp.h 7.4
SCCS-vsn: sys/netiso/clnp_debug.c 7.2
SCCS-vsn: sys/netiso/clnp_er.c 7.3
SCCS-vsn: sys/netiso/clnp_frag.c 7.3
SCCS-vsn: sys/netiso/clnp_input.c 7.4
SCCS-vsn: sys/netiso/clnp_options.c 7.3
SCCS-vsn: sys/netiso/clnp_output.c 7.3
SCCS-vsn: sys/netiso/clnp_raw.c 7.2
SCCS-vsn: sys/netiso/clnp_stat.h 7.2
SCCS-vsn: sys/netiso/clnp_subr.c 7.4
SCCS-vsn: sys/netiso/clnp_timer.c 7.2
SCCS-vsn: sys/netiso/cons.h 7.2
SCCS-vsn: sys/netiso/cons_pcb.h 7.2
SCCS-vsn: sys/netiso/eonvar.h 7.2
SCCS-vsn: sys/netiso/esis.c 7.2
SCCS-vsn: sys/netiso/esis.h 7.2
SCCS-vsn: sys/netiso/if_clnp.h 7.2
SCCS-vsn: sys/netiso/if_cons.c 7.2
SCCS-vsn: sys/netiso/if_eon.c 7.2
SCCS-vsn: sys/netiso/if_lpb.c 7.2
SCCS-vsn: sys/netiso/iso.c 7.2
SCCS-vsn: sys/netiso/iso.h 7.2
SCCS-vsn: sys/netiso/iso_addrs.c 7.2
SCCS-vsn: sys/netiso/iso_chksum.c 7.2
SCCS-vsn: sys/netiso/iso_errno.h 7.2
SCCS-vsn: sys/netiso/iso_map.h 7.2
SCCS-vsn: sys/netiso/iso_pcb.c 7.2
SCCS-vsn: sys/netiso/iso_pcb.h 7.2
SCCS-vsn: sys/netiso/iso_proto.c 7.2
SCCS-vsn: sys/netiso/iso_snpac.c 7.2
SCCS-vsn: sys/netiso/iso_snpac.h 7.2
SCCS-vsn: sys/netiso/iso_usrreq.c 7.2
SCCS-vsn: sys/netiso/iso_var.h 7.2
SCCS-vsn: sys/netiso/tp.trans 7.2
SCCS-vsn: sys/netiso/tp_astring.c 7.2
SCCS-vsn: sys/netiso/tp_clnp.h 7.2
SCCS-vsn: sys/netiso/tp_cons.c 7.2
SCCS-vsn: sys/netiso/tp_emit.c 7.2
SCCS-vsn: sys/netiso/tp_inet.c 7.2
SCCS-vsn: sys/netiso/tp_input.c 7.2
SCCS-vsn: sys/netiso/tp_ip.h 7.2
SCCS-vsn: sys/netiso/tp_iso.c 7.2
SCCS-vsn: sys/netiso/tp_meas.c 7.2
SCCS-vsn: sys/netiso/tp_meas.h 7.2
SCCS-vsn: sys/netiso/tp_output.c 7.2
SCCS-vsn: sys/netiso/tp_param.h 7.2
SCCS-vsn: sys/netiso/tp_pcb.c 7.2
SCCS-vsn: sys/netiso/tp_pcb.h 7.2
SCCS-vsn: sys/netiso/tp_seq.h 7.2
SCCS-vsn: sys/netiso/tp_sizes.c 7.2
SCCS-vsn: sys/netiso/tp_stat.h 7.2
SCCS-vsn: sys/netiso/tp_subr.c 7.2
SCCS-vsn: sys/netiso/tp_subr2.c 7.2
SCCS-vsn: sys/netiso/tp_timer.c 7.2
SCCS-vsn: sys/netiso/tp_timer.h 7.2
SCCS-vsn: sys/netiso/tp_tpdu.h 7.2
SCCS-vsn: sys/netiso/tp_trace.c 7.2
SCCS-vsn: sys/netiso/tp_trace.h 7.2
SCCS-vsn: sys/netiso/tp_user.h 7.2
SCCS-vsn: sys/netiso/tp_usrreq.c 7.2

+++ usr/src/sys/netiso/tp_trace.c
@@ -50,15 +50,15 @@ static char *rcsid = "$Header: tp_trace.c,v 5.3 88/11/18 17:29:14 nhall Exp $";
 #include "types.h"
 #include "time.h"
 
-#include "../netiso/tp_param.h"
-#include "../netiso/tp_timer.h"
-#include "../netiso/tp_stat.h"
-#include "../netiso/tp_param.h"
-#include "../netiso/tp_ip.h"
-#include "../netiso/tp_pcb.h"
-#include "../netiso/tp_tpdu.h"
-#include "../netiso/argo_debug.h"
-#include "../netiso/tp_trace.h"
+#include "tp_param.h"
+#include "tp_timer.h"
+#include "tp_stat.h"
+#include "tp_param.h"
+#include "tp_ip.h"
+#include "tp_pcb.h"
+#include "tp_tpdu.h"
+#include "argo_debug.h"
+#include "tp_trace.h"
 
 #ifdef TPPT
 static tp_seq = 0;
@@ -98,7 +98,7 @@ tpTrace(tpcb, event, arg, src, len, arg4, arg5)
 
 	case TPPTertpdu:
 		bcopy((caddr_t)src, (caddr_t)&tp->tpt_ertpdu,
-			MIN((int)len, sizeof(struct tp_Trace)));
+			(unsigned)MIN((int)len, sizeof(struct tp_Trace)));
 		break;
 
 	case TPPTusrreq:
@@ -106,7 +106,7 @@ tpTrace(tpcb, event, arg, src, len, arg4, arg5)
 
 		/* arg is a string */
 		bcopy((caddr_t)arg, (caddr_t)tp->tpt_str, 
-			MIN(1+strlen((caddr_t) arg), TPTRACE_STRLEN));
+			(unsigned)MIN(1+strlen((caddr_t) arg), TPTRACE_STRLEN));
 		tp->tpt_m2 = src; 
 		tp->tpt_m3 = len;
 		tp->tpt_m4 = arg4;
@@ -137,8 +137,8 @@ tpTrace(tpcb, event, arg, src, len, arg4, arg5)
 	case TPPTtpduin:
 	case TPPTtpduout:
 		tp->tpt_arg2 = arg4;
-		bcopy((caddr_t)src, (caddr_t)&tp->tpt_tpdu, MIN((int)len, 
-			sizeof(struct tp_Trace)));
+		bcopy((caddr_t)src, (caddr_t)&tp->tpt_tpdu,
+		      (unsigned)MIN((int)len, sizeof(struct tp_Trace)));
 		break;
 	}
 }
